+++ srml/support/src/storage/storage_items.rs
@@ -256,7 +256,7 @@ mod tests {
 	use crate::metadata::*;
 	use crate::metadata::StorageHasher;
 	use crate::rstd::marker::PhantomData;
-	use crate::storage::{StorageValue, StorageMap};
+	use crate::storage::StorageMap;
 
 	storage_items! {
 		Value: b"a" => u32;
