+++ srml/support/src/traits.rs
@@ -18,11 +18,13 @@
 //!
 //! NOTE: If you're looking for `parameter_types`, it has moved in to the top-level module.
 
-use crate::rstd::{prelude::*, result, marker::PhantomData, ops::Div};
-use crate::codec::{Codec, Encode, Decode};
+use rstd::{prelude::*, result, marker::PhantomData, ops::Div};
+use codec::{Codec, Encode, Decode};
 use primitives::u32_trait::Value as U32;
-use crate::sr_primitives::traits::{MaybeSerializeDebug, SimpleArithmetic, Saturating};
-use crate::sr_primitives::ConsensusEngineId;
+use sr_primitives::{
+	ConsensusEngineId,
+	traits::{MaybeSerializeDebug, SimpleArithmetic, Saturating},
+};
 
 /// Anything that can have a `::len()` method.
 pub trait Len {
